Added Spinner for Airline, Edited Add Trip Page
The Add Trip page is edited such that, we can select dates for Future
Trips and we cannot select a previous date. Also, The End Date of the
trip should be after the start date.

The Date is displayed on the button itself instead of separate Textview.

Changed Airline Trip Information from EditText to Spinner.

New String Array is created for Spinner Values.